requirements.txt
streamlit
yfinance
nltk
newsapi-python
plotly